refactor: cleanup type exprts
