+++ buf.gen.yaml
@@ -17,6 +17,11 @@ plugins:
       - compile_well_known_types
       - extern_path=.google.protobuf=::pbjson_types
       - file_descriptor_set=false
+      - type_attribute=.delta_sharing.internal.Resource.resource=#[derive(::strum::EnumDiscriminants)]
+      - type_attribute=.delta_sharing.internal.Resource.resource=#[strum_discriminants(name(ObjectLabel))]
+      - type_attribute=.delta_sharing.internal.Resource.resource=#[strum_discriminants(derive(::strum::AsRefStr\, ::strum::Display\, ::strum::EnumIter\, ::strum::EnumString)\, strum(serialize_all = "snake_case"))]
+      - type_attribute=.delta_sharing.internal.Resource.resource=#[strum_discriminants(cfg_attr(feature = "sqlx"\, derive(::sqlx::Type)))]
+      - type_attribute=.delta_sharing.internal.Resource.resource=#[strum_discriminants(cfg_attr(feature = "sqlx"\, sqlx(type_name = "object_label"\, rename_all = "snake_case")))]
   - remote: buf.build/community/neoeinstein-prost-serde:v0.3.1
     out: delta-sharing/common/src/gen
     opt:
