+++ delta-sharing/common/Cargo.toml
@@ -16,6 +16,7 @@ object_store = { workspace = true }
 pbjson = { workspace = true }
 prost = { workspace = true }
 serde = { workspace = true }
+strum = { workspace = true }
 thiserror = { workspace = true }
 tonic = { workspace = true, optional = true }
 tower = { workspace = true, features = ["make"], optional = true }
