+++ delta-sharing/common/src/models/properties.rs
